rename lj to ase_standard
